composition: extract preparation tests
